+++ src/components/index.ts
@@ -1,3 +1,4 @@
+export * from './HelpersEditor';
 export * from './Text';
 export * from './TextEditor';
 export * from './TextPanel';